# perf: avoid using delete operator