--- a/lib/hexo/post.js
+++ b/lib/hexo/post.js
@@ -208,7 +208,7 @@ class Post {
       // Create post
       Object.assign(data, yfmParse(content));
       data.content = data._content;
-      delete data._content;
+      data._content = undefined;
 
       return this.create(data, replace);
     }).then(post => {