--- a/lib/plugins/processor/asset.js
+++ b/lib/plugins/processor/asset.js
@@ -53,14 +53,14 @@ module.exports = ctx => {
       } else if (updated_option === 'date') {
         data.updated = data.date;
       } else if (updated_option === 'empty') {
-        delete data.updated;
+        data.updated = undefined;
       } else {
         data.updated = stats.mtime;
       }
 
       if (data.permalink) {
         data.path = data.permalink;
-        delete data.permalink;
+        data.permalink = undefined;
 
         if (data.path.endsWith('/')) {
           data.path += 'index';