--- a/lib/plugins/processor/post.js
+++ b/lib/plugins/processor/post.js
@@ -90,19 +90,19 @@ module.exports = ctx => {
       } else if (updated_option === 'date') {
         data.updated = data.date;
       } else if (updated_option === 'empty') {
-        delete data.updated;
+        data.updated = undefined;
       } else {
         data.updated = stats.mtime;
       }
 
       if (data.category && !data.categories) {
         data.categories = data.category;
-        delete data.category;
+        data.category = undefined;
       }
 
       if (data.tag && !data.tags) {
         data.tags = data.tag;
-        delete data.tag;
+        data.tag = undefined;
       }
 
       categories = data.categories || [];
@@ -113,7 +113,7 @@ module.exports = ctx => {
 
       if (data.photo && !data.photos) {
         data.photos = data.photo;
-        delete data.photo;
+        data.photo = undefined;
       }
 
       if (data.photos && !Array.isArray(data.photos)) {
@@ -126,7 +126,7 @@ module.exports = ctx => {
 
       if (data.permalink) {
         data.__permalink = data.permalink;
-        delete data.permalink;
+        data.permalink = undefined;
       }
 
       // FIXME: Data may be inserted when reading files. Load it again to prevent