--- a/lib/theme/index.js
+++ b/lib/theme/index.js
@@ -72,7 +72,7 @@ class Theme extends Box {
 
     if (!views) return;
 
-    delete views[ext];
+    views[ext] = undefined;
   }
 }