<!--
Thank you for creating a pull request to contribute to the Hexo code! Before you open the request please answer the following questions to help it be more easily integrated. Please check the boxes "[ ]" with "[x]" when done too.
-->

## What does it do?

https://v8.dev/blog/fast-properties

In short, the `delete` operator has a potential performance impact (related with v8 hidden class, and `delete` could cause v8 to degrade object process). By replacing `delete` operator it could gain performance.

## Screenshots

Benchmark result:

**master branch**

```
Cold processing
┌──────────────────────────────┬───────────────┐
│           (index)            │ Cost time (s) │
├──────────────────────────────┼───────────────┤
│ Load Plugin/Scripts/Database │    '0.41s'    │
│        Process Source        │    '3.60s'    │
│         Render Posts         │    '3.36s'    │
│         Render Files         │    '7.27s'    │
│        Save Database         │    '0.35s'    │
│          Total time          │   '14.99s'    │
└──────────────────────────────┴───────────────┘
Hot processing
┌──────────────────────────────┬───────────────┐
│           (index)            │ Cost time (s) │
├──────────────────────────────┼───────────────┤
│ Load Plugin/Scripts/Database │    '1.20s'    │
│        Process Source        │    '1.41s'    │
│         Render Files         │    '7.23s'    │
│        Save Database         │    '0.35s'    │
│          Total time          │   '10.20s'    │
└──────────────────────────────┴───────────────┘
Another Cold processing
┌──────────────────────────────┬───────────────┐
│           (index)            │ Cost time (s) │
├──────────────────────────────┼───────────────┤
│ Load Plugin/Scripts/Database │    '0.41s'    │
│        Process Source        │    '3.43s'    │
│         Render Posts         │    '3.31s'    │
│         Render Files         │    '7.38s'    │
│        Save Database         │    '0.34s'    │
│          Total time          │   '14.87s'    │
└──────────────────────────────┴───────────────┘
```

After the PR:

```
Cold processing
Cold processing
┌──────────────────────────────┬───────────────┐
│           (index)            │ Cost time (s) │
├──────────────────────────────┼───────────────┤
│ Load Plugin/Scripts/Database │    '0.36s'    │
│        Process Source        │    '3.32s'    │
│         Render Posts         │    '3.22s'    │
│         Render Files         │    '6.94s'    │
│        Save Database         │    '0.33s'    │
│          Total time          │   '14.17s'    │
└──────────────────────────────┴───────────────┘
Hot processing
┌──────────────────────────────┬───────────────┐
│           (index)            │ Cost time (s) │
├──────────────────────────────┼───────────────┤
│ Load Plugin/Scripts/Database │    '1.07s'    │
│        Process Source        │    '1.38s'    │
│         Render Files         │    '6.80s'    │
│        Save Database         │    '0.34s'    │
│          Total time          │    '9.59s'    │
└──────────────────────────────┴───────────────┘
Another Cold processing
┌──────────────────────────────┬───────────────┐
│           (index)            │ Cost time (s) │
├──────────────────────────────┼───────────────┤
│ Load Plugin/Scripts/Database │    '0.38s'    │
│        Process Source        │    '3.32s'    │
│         Render Posts         │    '3.14s'    │
│         Render Files         │    '6.82s'    │
│        Save Database         │    '0.35s'    │
│          Total time          │   '14.01s'    │
└──────────────────────────────┴───────────────┘
```

Performance improved by 7% after the PR.

## Pull request tasks

- [ ] Add test cases for the changes.
- [x] Passed the CI test.